changes 2019.12.04
(1)    please use the attached photos for the two chocolate options instead of the earlier ones.

(2)    In addition to CH (chocolate), HL (hard lottery) and EA (easy lottery), there will be a fourth decision situation. It will also involve two options – pics are attached. It is a time preference decision situation (TP). I think the pics still need cutting – do you know how to do this with PNG?

(3)    Please replace the block I question with the text and list as in the word doc attached. The list should be displayed as it is currently, so NOT with 200 rows. Participants should only be able to enter a value between 0 and 200 in the open text field.