changes
could you please remove the white space around the chocolate bars (sorry, I should have done this before sending it!)

please use the TP with the 3 lines, it does not show up right in firefox (one option is outside of the screen)

block I and II: participants should only be able to type in a value, not use the click up/down buttons.

Block II: could you please put the options a bit closer towards the middle

Please only show the final payoff and call it “The computer has now randomly determined your payoff for this experiment. Your payoff is:”

We will not need the questions about the person screen, we will do this on paper.